Move blocked impl. to separate file
