--- src/svd/two-sided/main.cpp
@@ -1,1 +1,0 @@
-int main(void) { return 0; }
